requirements.txt
flask>=3.0
flask-cors>=4.0
openai>=1.0
python-dotenv>=1.0
imageio-ffmpeg>=0.5
